apps/baseten-livekit-agent/pyproject.toml
[project]
name = "sanne-baseten-livekit-agent"
version = "0.1.0"
description = "Baseten-deployable LiveKit worker for the Sanne demo"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "livekit-agents[google,mistralai,openai,silero]~=1.5",
    "livekit-plugins-noise-cancellation>=0.2.0",
    "numpy>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
